move ngram normalizer
