+++ src/fuzzy-searchers/index.ts
@@ -1,6 +1,7 @@
 export { FuzzySearcher as FuzzySearcherImpl } from './fuzzy-searcher.js';
 export { FuzzySearchConfig } from './fuzzy-search-config.js';
 export { InvertedIndex } from './inverted-index.js';
+export { NgramNormalizer } from './ngram-normalizer.js';
 export { NgramComputer } from './ngram-computer.js';
 export { QualityComputer } from './quality-computer.js';
 export { TermIds } from './term-ids.js';
